+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added more tooltips for date filters to inform about the possibility to use "*" as wildcards
 - Fixed a CSS issue where tooltips weren't shown on top of other content
 - Moved the feedback button from the Settings to the header/sidebar
+- Fixed an issue where the Tool and Media type dropdown, Username field and (Re)Load button would vanish during window resizing
 
 ## 2025-01-03
 - Add "Remove all values" buttons on Tagify instances